+++ test/data/safety/content/test/pin-with-imports.js
@@ -1,0 +1,5 @@
+import nock from "nock";
+
+process.env.NODE_CONFIG_ENV = "test";
+
+nock.disableNetConnect();