+++ test/data/safety/content/test/pin-with-require.js
@@ -1,0 +1,5 @@
+const logger = require("../../../../../lib/logger.js");
+
+process.env.NODE_CONFIG_ENV = "test";
+
+module.exports = logger;